+++ packages/terra-avatar/src/terra-dev-site/doc/example/facility/FacilitySize.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { Facility } from 'terra-avatar';
-import image from '../../../assets/200x133.jpg';
-
-const AvatarSize = () => (
-  <Facility image={image} alt="Thailand" size="2em" />
-);
-
-export default AvatarSize;