+++ packages/terra-avatar/src/terra-dev-site/test/avatar/avatar/TwoInitialsAvatar.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Avatar from '../../../../index';
-
-export default () => <Avatar id="two-initials-avatar" initials="JSS" alt="Joe S Shane" color="two" />;