+++ packages/terra-avatar/src/terra-dev-site/test/avatar/facility/ImageFacility.test.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import { Facility } from '../../../../index';
-import exampleAvatarImage from '../../../assets/150x150.jpg';
-
-export default () => <Facility id="image-facility" image={exampleAvatarImage} alt="facility" />;