+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/avatar/IsDeceasedInitialsAvatar.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Avatar from 'terra-avatar';
+
+export default () => <Avatar id="is-deceased-initials-avatar" alt="user" isDeceased initials="AA" />;