+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/avatar/TwoInitialsAvatar.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Avatar from 'terra-avatar';
+
+export default () => <Avatar id="two-initials-avatar" initials="JSS" alt="Joe S Shane" color="two" />;