+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/facility/ImageFacility.test.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import { Facility } from 'terra-avatar';
+import exampleAvatarImage from '../../../doc/avatar/assets/150x150.jpg';
+
+export default () => <Facility id="image-facility" image={exampleAvatarImage} alt="facility" />;