+++ packages/terra-core-docs/src/terra-dev-site/test/avatar/generic/Provider.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Generic } from 'terra-avatar';
+
+export default () => <Generic variant="provider" id="provider" alt="provider" color="seven" />;