+++ packages/terra-avatar/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added package json lint test.
 
+* Changed
+  * migrate doc to terra-core-docs
+
 ## 3.13.0 - (September 21, 2021)
 
 * Changed